moved search suggestion to components
